Delay delete of detached hosts based on annotation metadata
This adds optional metadata to the detached annotation which allows a
user to delay the deletion of a host.

This is useful in a case where another controller wants the BMH detached
to avoid extra operations on the host, but also wants to take advantage
of deprovisioning. Such a controller would set the detached annotation
delete action to "delay" and also set its own finalizer on the BMH.
Then, when the BMH is marked for deletion BMO would wait until the
detached annotation is removed giving the other controller time to take
whatever actions are needed (setting cleaning, changing images, etc)